[MOVE] 공통 에러 이관
