+++ app/common/src/main/java/com/chinda/common/exception/HttpException.java
@@ -1,4 +1,4 @@
-package com.common.common.exception.model;
+package com.chinda.common.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
