+++ app/common/src/main/java/com/chinda/common/exception/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.common.common.exception.dto;
+package com.chinda.common.exception.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
